Time: 19 ms (29.43%) | Memory: 19.9 MB (8.83%) - LeetSync

+++ 169-majority-element/majority-element.cpp
@@ -11,8 +11,8 @@ public:
         //         return i.first;
         // }
         // return -1;
-        
+
         sort(nums.begin(),nums.end());
-        return nums[(int)nums.size()/2];  
+        return nums[nums.size()/2];  
     }
 };
